# Backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: SocietyEase API
  description: |
    API specification for the SocietyEase backend (Flask). Authentication uses JWT Bearer tokens. Each API endpoint is mapped to its corresponding user story using the x-user-story extension for traceability between the implementation and the project requirements.

    ## User story mapping

    Every operation carries an `x-user-story` extension. The stories are:

    | ID | User story |
    |---|---|
    | **US-01** | As a **Treasurer**, I want to generate and track maintenance invoices for every flat, so that dues are billed consistently and I can see what is outstanding. |
    | **US-02** | As a **Resident**, I want to raise a complaint about a problem in my flat or the building, so that the society has a record of it and someone is accountable for fixing it. |
    | **US-03** | As a **Resident**, I want to track the status of my complaint from open to closed, so that I know whether it is being worked on without having to chase anyone. |
    | **US-04** | As a **Association manager**, I want to assign a complaint to a maintenance worker, so that the right person is responsible and the job reaches their queue. |
    | **US-05** | As a **Owner**, I want to see the invoices raised against the flat I own, so that I can keep track of what my tenant or I owe the society. |
...

    The same table is maintained in `docs/USER_STORIES.md`.
  version: 4.1.0
servers:
- url: http://localhost:5000
...
paths:
...
  /api/members/apartments:
    get:
      tags:
      - Members
      summary: List apartments
      description: |-
        Returns every apartment/flat registered in the society. **User Story — US-09**

        Access: any authenticated, active user.
      x-user-story: US-09
      security:
      - bearerAuth: []
      responses:
        '200':
          description: List of apartments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Apartment'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_get_apartments
      x-required-role: any authenticated, active user
    post:
      tags:
      - Members
      summary: Add apartment
      description: |-
        Registers a new apartment/flat. `flat_number` must be unique across the society. **User Story — US-09**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-09
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                flat_number:
                  type: string
                block:
                  type: string
                floor:
                  type: integer
              required:
              - flat_number
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '201':
          description: Apartment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Apartment'
        '400':
          description: flat_number is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: An apartment with this flat_number already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_add_apartment
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
  /api/members/apartments/{apt_id}:
    put:
      tags:
      - Members
      summary: Update apartment
      description: |-
        Updates the block and/or floor of an existing apartment. **User Story — US-09**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-09
      security:
      - bearerAuth: []
      parameters:
      - name: apt_id
        in: path
        required: true
        schema:
          type: integer
        description: Apartment (flat) id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                block:
                  type: string
                floor:
                  type: integer
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '200':
          description: Apartment updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Apartment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No apartment with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_update_apartment
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
    delete:
      tags:
      - Members
      summary: Delete apartment
      description: |-
        Removes a flat. **Refused with 409 if the flat still has residents or invoices**, so deleting a flat can never silently destroy its residents, invoices, payments or complaints. Detach or remove those records first. **User Story — US-09**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-09
      security:
      - bearerAuth: []
      parameters:
      - name: apt_id
        in: path
        required: true
        schema:
          type: integer
        description: Apartment (flat) id.
      responses:
        '200':
          description: Apartment deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No apartment with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_delete_apartment
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
  /api/members/:
    get:
      tags:
      - Members
      summary: List members
      description: |-
        Lists every resident with their flat, role and contact details, optionally narrowed by any combination of the query filters below. **Admin-only** (ADMIN, SYSTEM_ADMIN, TREASURER, COMMITTEE_MEMBER) — this is society-wide personal data. **User Story — US-09, US-18**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story:
      - US-09
      - US-18
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
        description: Free-text search across name, email, phone and flat number.
      - name: role
        in: query
        schema:
          type: string
        description: Exact role match, e.g. TENANT, OWNER, WORKER.
      - name: is_owner
        in: query
        schema:
          type: boolean
        description: 'true = owners only, false = tenants only.'
      - name: is_active
        in: query
        schema:
          type: boolean
      - name: apartment_id
        in: query
        schema:
          type: integer
      - name: block
        in: query
        schema:
          type: string
      - name: floor
        in: query
        schema:
          type: integer
      - name: has_parking
        in: query
        schema:
          type: boolean
        description: Whether the member's flat currently has an occupied or reserved parking slot.
      responses:
        '200':
          description: List of members, optionally filtered.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_get_members
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
    post:
      tags:
      - Members
      summary: Add member
      description: |-
        Creates a new user account and links it to an apartment as a resident, in one call. Intended for association-manager use. **User Story — US-09**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-09
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                apartment_id:
                  type: integer
                phone:
                  type: string
                is_owner:
                  type: boolean
                move_in_date:
                  type: string
              required:
              - name
              - email
              - password
              - role
              - apartment_id
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '201':
          description: Member (user + resident record) created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resident'
        '400':
          description: A required field is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: apartment_id does not match an existing apartment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A user with this email is already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_add_member
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
  /api/members/export:
    get:
      tags:
      - Members
      summary: Export members as CSV
      description: |-
        Same role scoping and query filters as GET /api/members/, returned as a downloadable CSV instead of JSON. **User Story — US-19**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-19
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: role
        in: query
        schema:
          type: string
      - name: is_owner
        in: query
        schema:
          type: boolean
      - name: is_active
        in: query
        schema:
          type: boolean
      - name: apartment_id
        in: query
        schema:
          type: integer
      - name: block
        in: query
        schema:
          type: string
      - name: floor
        in: query
        schema:
          type: integer
      - name: has_parking
        in: query
        schema:
          type: boolean
      responses:
        '200':
          description: CSV file of the filtered member list.
          content:
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_export_members
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
  /api/members/{resident_id}:
    get:
      tags:
      - Members
      summary: Get member detail
      description: |-
        Returns a single resident's profile and flat details. **User Story — US-09**

        Access: any authenticated, active user.
      x-user-story: US-09
      security:
      - bearerAuth: []
      parameters:
      - name: resident_id
        in: path
        required: true
        schema:
          type: integer
        description: Resident record id — **not** the users.id.
      responses:
        '200':
          description: Member detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resident'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No resident with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_get_member
      x-required-role: any authenticated, active user
    put:
      tags:
      - Members
      summary: Update member
      description: |-
        Updates a resident's name, phone, ownership flag, and move-in/move-out dates. **User Story — US-09**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-09
      security:
      - bearerAuth: []
      parameters:
      - name: resident_id
        in: path
        required: true
        schema:
          type: integer
        description: Resident record id — **not** the users.id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                is_owner:
                  type: boolean
                move_in_date:
                  type: string
                move_out_date:
                  type: string
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '200':
          description: Member updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No resident with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_update_member
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
    delete:
      tags:
      - Members
      summary: Deactivate member
      description: |-
        Soft-deletes a member by setting the linked user account's is_active flag to false. The resident record itself is not deleted. **User Story — US-09**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-09
      security:
      - bearerAuth: []
      parameters:
      - name: resident_id
        in: path
        required: true
        schema:
          type: integer
        description: Resident record id — **not** the users.id.
      responses:
        '200':
          description: Member deactivated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No resident with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: members_deactivate_member
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
...
  /api/members/workers:
    get:
      tags:
      - Members
      summary: List maintenance workers
      description: |-
        Returns every active user whose role is WORKER, for populating the worker picker when assigning a complaint. `id` is the **users.id**, which is what `PUT /api/complaints/{cid}/assign` expects as `worker_id` (not the resident id).

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-04
      security:
      - bearerAuth: []
      responses:
        '200':
          description: The list of active maintenance workers.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: users.id — use as worker_id
                    name:
                      type: string
                    email:
                      type: string
                      format: email
              example:
              - id: 6
                name: Ramesh Worker
                email: worker@apt.com
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: members_get_workers
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
      - error
...
    Apartment:
      type: object
      properties:
        id:
          type: integer
        block:
          type: string
          nullable: true
        flat_number:
          type: string
        floor:
          type: integer
      required:
      - id
      - flat_number
...
  responses:
...
    Forbidden:
      description: 'The caller is authenticated but not allowed: either the account has been deactivated,
        or the user''s role is not permitted to perform this action, or the resource belongs to another
        resident.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: You are not allowed to perform this action
...
    ServerError:
      description: Unexpected server error. All errors are returned as JSON; the API never returns an
        HTML error page.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
...